DIBR-4056 Fixed the improper resource access if it's not allocated (#7)
This fix is for meta api and proxy access - now only the allocated application resource will be able to get the metadata, access to proxy and the resource-by-ip. Otherwise it's getting weird if we are able to get or modify the resource parameters from the machine that's not the actual resource.