Added AWS driver (#24)
The change contains a couple of minor improvements along with AWS driver which doesn't require any additional dependencies and actually could provision a good instances in the prepared environment.

It not just adds AWS as a driver, but also adds snapshot API interface (which is implemented now just for AWS but potentially can work with VMX and other drivers too). The idea behind it is to create the pre-filled workspaces for cache or to save the state of the instance if it needs some additional debugging elsewhere. It's not implemented properly and will work only on the node that has the AWS account, later need to be rewritten to work as expected in clustered environment.