Migrated to SQlite, added CI & tests (#19)
This PR is intended to fix a multiple issues with DQlite. It moves to the SQlite and seriously decreases the amount of dependencies to build, uses port of SQlite to go and enforces the cluster protection. Now the nodes will talk using websocket and the functionality of it soon will be added - for now quite rudimentary.

Also the change adds actual CI & simple integration tests to the repository.

The change introduces a new way to operate a cluster and will not be compatible with the previously created raft files - hopefully for the best. API also was changed quite dramatically - all the sequential ID's become UID's to operate properly on cluster level, but overall this should not be so big deal to migrate to this change.

* Switched to sqlite, improved cert security and added cluster cert auth
* Added Cluster client which is connecting to another node
* Migrated to non-sequential database UIDs
* Migrated to oapi-codegen v1.12.2
* Added none driver & simple automated integration test
* Added github actions CI workflow & fixed aws driver docs
* Fixed CodeQL complains & issues