Redesigned the Labels to support multidefinition operation (#37)
The labels now can contain multiple definitions and the right one will be chosen by the lowest number in Available field of the Vote. If Available is -1 than the node can't serve any of the Label definitions. Also Resource contains new fields to easily find which of the label definitions was chosen.

The drivers also got some redesign to store Resources separately to be properly operated outside of the driver. Previously we had to go to the driver definition to figure out what kind of resources it's consuming, but now it's all parsed automatically, only driver specific Options now handled by the driver itself.

Unfortunately this change is breaking and incompatible with the previous versions due to rewrite of the interfaces.