This one I don't like, but it's the only option because of AWS architecture
When the host is changing state from Pending to Available - it doesn't
mean that it's actually available for the instance (sic). The allocation
can easily fail because distributed architecture of AWS doesn't allow a
certain knowledge of the state and there is no such a thing as priority
in syncing state update messages to update API & Dashboard last, when
the region internally is updated. So we could easily see an error even
if the automation is perfect.

So I'm adding automatic retries to all the AWS calls with pain in my
heart. Maybe someday will be a better way of AWS will fix the issue...