Added simple dedicated hosts simulator to predict quota and price per month (#86)
This change adds the simulator, because it's quite complicated to figure out how much you need and how much will you pay for that. So simulator intakes the workload events with usage duration and passes it through the scrutiny of dedicated host allocation and acts just like the actual Fish AWS driver when it comes to manage the restricted resources.

For sure simulator gives not 100% accurate results, because there are a ton of other variables (like variable scrubbing time and other delays), but it gets close enough to tell how much you will pay for the luxury with enabled scrubbing optimization.