Fixed a nasty bug with stored hashes of passwords
Hash is generated with random salt by default and if we don't know
salt (our case) - we can't pull from DB and remove the access quickly.
So instead I store just hash there with empty salt. This works, because
the access is temporary and one-shot, so technically no issues, but
if we will change that at some point - this mechanism is needed to be
reworked.