Ok, finally figured out the last issue we had with the tests
Apparently the Mac & Lin network layers (or most probably ssh clients of
different versions) are handling the input a bit differently. If you will
easily get away with just buffer as input on Mac and SSH (9.6) client
will not close the src->dst channel - it's not the case with Linux SSH
(9.2p1) client: it will do that and you would never see any response from
the server. So instead we now emulating the ssh terminal input as well
and not closing the input until it's necessary.

Also I found a wrong src->dst messages, so flipped them over and grately
improved the ssh testing where it knocks the mock SSHD first and then
retries the same with proxy. And last thing - found a couple of lost
errors because serve is running as a goroutine and overall increased
verbosity of the mock sshd, so should be much easier to debug now.