AF-102 Reorganized proxies in driver flavor and changed drivers config structure (#103)
In order to keep things organized and prepare for the future changes I had to rewrite the way database exists withing Fish core. Separating the database into the module made it possible to use database functions in drivers, so proxies now a part of usual drivers config flow - by default they are enabled, but it's easy to disable them by specifying empty map in config.

**WARNING:** Backward-incompatible in configuration, `drivers` in fish's config.yml have to be altered to fit the new schema.