AF-113 Increased AWS stop timeout to 30 mins and made snapshot fail if instance is not stopped (#115)
Fixes the issue with broken snapshot as a result of the task. The consistence is more important then no matter what prepare a snapshot, especially if it's a broken one.

Also partially fixed issue with openssh - now test tty client waits a second for shell response before beginning input - that solves some issues with too early input, but not all of them - for example it seems proxy closes the channel too soon and don't receive "exit", but that's an issue for future me.